requirements.txt
aiogram==3.6.0
python-dotenv>=1.0.1
asyncpg>=0.29
SQLAlchemy[asyncio]>=2.0
psycopg[binary]>=3.1
openai>=1.30
ffmpeg-python>=0.2
pydub>=0.25
email-validator>=2.0